refactor: organize files into features
