+++ src/App.tsx
@@ -1,9 +1,4 @@
-import { Shell } from "@/components/Shell";
-import { ChatPage } from "@/pages/ChatPage";
+import { Layout } from "@/features/layout/ui/Layout";
 import { FC } from "react";
 
-export const App: FC = () => (
-  <Shell>
-    <ChatPage />
-  </Shell>
-);
+export const App: FC = () => <Layout />;
