--- src/components/Shell/index.ts
@@ -1,1 +1,0 @@
-export * from "./Shell";
