+++ src/features/chat/ui/Chat/index.ts
@@ -1,0 +1,1 @@
+export * from "./Chat";
