+++ src/features/layout/ui/Layout/Layout.tsx
@@ -1,0 +1,12 @@
+import { ChatPage } from "@/features/chat/ui/Chat";
+import { AppShell } from "@mantine/core";
+import { FC, PropsWithChildren } from "react";
+import classes from "./Layout.module.css";
+
+export const Layout: FC<PropsWithChildren> = () => (
+  <AppShell>
+    <AppShell.Main className={classes.main}>
+      <ChatPage />
+    </AppShell.Main>
+  </AppShell>
+);
