--- src/pages/ChatPage/index.ts
@@ -1,1 +1,0 @@
-export * from "./ChatPage";
